Added current working versions
The Scraping_in_R file has a duplicate set of pulls to get info from proteintech about their antibodies.  It still needs to be modified to pull reactivity data reliably, as proteintech's sites are apparently not all uniform when they are generated.  I would also like to add functionality for other major antibody manufacturers (e.g. abcam, sigma, etc.), and, if possible, I would like to have it use website searches so that you only have to enter in catalog numbers, not the whole url for the antibody.